chore: fix review comments
