+++ app/client/packages/design-system/widgets/src/components/Checkbox/src/index.ts
@@ -1,2 +1,2 @@
 export { Checkbox } from "./Checkbox";
-export type { CheckboxProps } from "./Checkbox";
+export type { CheckboxProps } from "./types";
