+++ app/client/packages/design-system/widgets/src/components/Checkbox/src/types.ts
@@ -1,0 +1,6 @@
+import type { CheckboxProps as HeadlessCheckboxProps } from "react-aria-components";
+import type { POSITION } from "../../../shared";
+
+export interface CheckboxProps extends HeadlessCheckboxProps {
+  labelPosition?: keyof typeof POSITION;
+}
