+++ app/client/packages/design-system/widgets/src/components/RadioGroup/src/RadioGroup.tsx
@@ -1,6 +1,5 @@
 import React, { forwardRef, useRef } from "react";
 import { RadioGroup as HeadlessRadioGroup, Radio } from "react-aria-components";
-import type { ORIENTATION } from "@design-system/widgets";
 import {
   Label,
   Flex,
@@ -10,39 +9,7 @@ import {
 } from "@design-system/widgets";
 import styles from "./styles.module.css";
 import type { ForwardedRef } from "react";
-import type { RadioGroupProps as HeadlessRadioGroupProps } from "react-aria-components";
-
-interface RadioGroupItemProps {
-  value: string;
-  label?: string;
-  isSelected?: boolean;
-  isDisabled?: boolean;
-  index?: number;
-}
-
-export interface RadioGroupProps extends HeadlessRadioGroupProps {
-  /**
-   * A ContextualHelp element to place next to the label.
-   */
-  contextualHelp?: string;
-  /**
-   * The content to display as the label.
-   */
-  label?: string;
-  /**
-   * Radio that belong to this group.
-   */
-  items: RadioGroupItemProps[];
-  /**
-   * The axis the checkboxes should align with.
-   * @default 'horizontal'
-   */
-  orientation?: keyof typeof ORIENTATION;
-  /**
-   * An error message for the field.
-   */
-  errorMessage?: string;
-}
+import type { RadioGroupProps } from "./types";
 
 const _RadioGroup = (
   props: RadioGroupProps,
