+++ app/client/packages/design-system/widgets/src/components/RadioGroup/src/index.ts
@@ -1,2 +1,2 @@
 export { RadioGroup } from "./RadioGroup";
-export type { RadioGroupProps } from "./RadioGroup";
+export type { RadioGroupProps } from "./types";
